Combined Terminal and Data Product logic
