--- .changeset/calm-icons-wish.md
@@ -1,6 +1,0 @@
----
-'@finos/legend-application-marketplace-deployment': patch
-'@finos/legend-application-marketplace': patch
----
-
-Redesign Marketplace Lakehouse home page
