--- .changeset/little-clouds-grow.md
@@ -1,7 +1,0 @@
----
-'@finos/legend-application-data-cube': patch
-'@finos/legend-application': patch
-'@finos/legend-shared': patch
----
-
-Add sanitizeParametersInsteadOfUrl parameter to getQueryParameterValue function
