+++ packages/legend-application-marketplace/src/pages/Lakehouse/dataProduct/LakehouseSDLCDataProduct.tsx
@@ -30,7 +30,7 @@ import {
 } from '../../../__lib__/LegendMarketplaceNavigation.js';
 import { useParams } from '@finos/legend-application/browser';
 import { guaranteeNonNullable } from '@finos/legend-shared';
-import { DataProductViewer } from './DataProductViewer.js';
+import { ProductViewer } from './ProductViewer.js';
 import { LEGEND_APPLICATION_COLOR_THEME } from '@finos/legend-application';
 import { LegendMarketplacePage } from '../../LegendMarketplacePage.js';
 
@@ -67,8 +67,8 @@ export const LakehouseSDLCDataProduct = withMarketplaceLakehouseStore(
           <CubesLoadingIndicatorIcon />
         </CubesLoadingIndicator>
         {marketPlaceStore.dataProductViewer && (
-          <DataProductViewer
-            dataSpaceViewerState={marketPlaceStore.dataProductViewer}
+          <ProductViewer
+            productViewerState={marketPlaceStore.dataProductViewer}
           />
         )}
       </LegendMarketplacePage>
