+++ packages/legend-application-marketplace/style/components/_data-product-viewer.scss
@@ -150,23 +150,6 @@
         background: var(--color-dark-grey-250);
       }
     }
-
-    &__actions-selector {
-      @include flexCenter;
-
-      width: 3.8rem;
-      height: 3.8rem;
-      border-left: 0.1rem solid var(--color-dark-grey-250);
-
-      svg {
-        font-size: 1.6rem;
-        color: var(--color-dark-grey-400);
-      }
-
-      &:hover svg {
-        color: var(--color-light-grey-400);
-      }
-    }
   }
 
   &__body {
@@ -200,114 +183,6 @@
     width: 100%;
     padding: 10rem 4rem 1rem;
     box-shadow: var(--color-dark-shade-280) 0 0.1rem 0.5rem 0;
-
-    &__access-button {
-      width: 134px;
-      height: 30px;
-      border-radius: 4px;
-      background-color: var(--color-hclight-blue-60);
-      color: var(--color-hclight-white);
-      border: none;
-      cursor: pointer;
-      font-size: 14px;
-      font-weight: 500;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      transition: all 0.2s ease;
-      outline: none;
-
-      &:hover {
-        opacity: 0.9;
-      }
-
-      &--pending {
-        background-color: #ef6c00;
-      }
-
-      &--entitled {
-        background-color: #2e7d32;
-        padding: 0 12px;
-        justify-content: space-between;
-
-        &--dropdown {
-          width: 195px;
-          padding: 8px 12px;
-          border: 1px;
-          border-radius: 4px;
-          background-color: white;
-          text-align: left;
-          cursor: pointer;
-          font-size: 14px;
-          color: #333;
-          z-index: 1000;
-
-          // Need above for conflicting style rules
-          box-shadow: var(--color-dark-shade-100) 0 0.1rem 0.5rem 0;
-
-          &:hover {
-            background-color: #f5f5f5;
-            box-shadow: var(--color-dark-shade-200) 0 0.1rem 0.5rem 0;
-          }
-        }
-      }
-
-      &:disabled {
-        opacity: 0.5;
-        cursor: not-allowed;
-      }
-    }
-
-    &__support-section {
-      &__container {
-        padding-top: 70px;
-        padding-bottom: 20px;
-
-        &__div {
-          display: grid;
-          gap: 20px;
-          grid-template-columns: repeat(2, 1fr);
-          grid-template-rows: repeat(3, auto);
-          grid-auto-flow: column;
-          width: 50%;
-          padding: 10px;
-        }
-      }
-
-      &__heading {
-        font-size: 20px;
-        font-weight: 500;
-      }
-
-      &__link {
-        text-decoration: none;
-        font-size: 14px;
-        color: #1b64a7;
-        font-weight: 500;
-        // word-break: break-word;
-
-        &--item {
-          display: flex;
-          flex-direction: row;
-          align-items: center;
-          text-align: left;
-          white-space: nowrap;
-        }
-
-        &--circle {
-          width: 36px;
-          height: 36px;
-          border-radius: 50%;
-          background-color: #2c2c2c;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          margin-right: 12px;
-          overflow: hidden;
-          flex-shrink: 0;
-        }
-      }
-    }
 
     &__table {
       width: 100%;
